requirements.txt
python-telegram-bot==20.3
beautifulsoup4
requests
